Random Color RecyclerView

+++ app/src/main/java/dead/code/note/ui/NoteFragment.kt
@@ -1,5 +1,6 @@
 package dead.code.note.ui
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -8,6 +9,7 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
+import dead.code.note.R
 import dead.code.note.data.Note
 import dead.code.note.databinding.FragmentNoteBinding
 import dead.code.note.viewmodel.NotesViewModel
@@ -27,8 +29,7 @@ class NoteFragment : Fragment() {
     }
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ): View {
         binding = FragmentNoteBinding.inflate(inflater)
         return binding.root
@@ -51,8 +52,7 @@ class NoteFragment : Fragment() {
                 val noteText = edNote.text.toString()
                 Note(id, noteTitle, noteText).also { note ->
                     if (noteTitle.isEmpty() && noteText.isEmpty()) {
-                        Toast.makeText(context, "All fields must be filled", Toast.LENGTH_LONG)
-                            .show()
+                        Toast.makeText(context, "All fields must be filled", Toast.LENGTH_LONG).show()
                         return@setOnClickListener
                     }
                     viewModel.upsertNote(note)
@@ -66,13 +66,31 @@ class NoteFragment : Fragment() {
                 val noteID = args.note!!.noteId
                 val noteTitle = edTitle.text.toString()
                 val noteText = edNote.text.toString()
-                Note(noteID, noteTitle, noteText).also {
-                    viewModel.deleteNotes(it)
-                    findNavController().navigateUp()
-                }
-
+                alertDelete(noteID, noteTitle, noteText)
             }
         }
 
+
     }
+
+    private fun alertDelete(id: Int, title: String, text: String): AlertDialog {
+        val alertDelete = AlertDialog.Builder(requireActivity()).apply {
+            setIcon(R.drawable.ic_delete)
+            setMessage("Do you want delete it")
+            setPositiveButton("yes") { _, _ ->
+                Note(id, title, text).also {
+                    viewModel.deleteNotes(it)
+                    findNavController().navigateUp()
+                }
+            }
+            setNegativeButton("No") { d, _ ->
+                d.dismiss()
+                findNavController().navigateUp()
+            }
+            create()
+        }
+        return alertDelete.show()
+    }
+
+
 }
